@@ -1,25 +1,26 @@
 {
   "name": "stif-boiv",
   "dependencies": {
-    "@rails/webpacker": "3.0.2",
-    "babel-jest": "21.2.0",
-    "babel-polyfill": "6.16.0",
-    "babel-preset-es2015": "6.18.0",
+    "@rails/webpacker": "3.2.0",
+    "babel-jest": "22.0.4",
+    "babel-polyfill": "6.26.0",
+    "babel-preset-es2015": "6.24.1",
     "babel-preset-react": "6.24.1",
-    "babelify": "7.3.0",
+    "babelify": "8.0.0",
     "bootstrap": "3",
-    "coffeescript": "1.12.7",
+    "coffeescript": "2.1.0",
     "jquery": "3.2.1",
     "lodash": "4.17.4",
-    "promise-polyfill": "6.0.2",
-    "react": "15.3.2",
-    "react-dom": "15.3.2",
-    "react-redux": "4.4.5",
-    "react-select2": "4.0.3",
-    "redux": "3.6.0",
-    "redux-logger": "2.7.4",
+    "promise-polyfill": "7.0.0",
+    "prop-types": "^15.6.0",
+    "react": "16.2.0",
+    "react-dom": "16.2.0",
+    "react-redux": "5.0.6",
+    "react-select2-wrapper": "^1.0.4-beta5",
+    "redux": "3.7.2",
+    "redux-logger": "3.0.6",
     "redux-promise": "0.5.3",
-    "redux-thunk": "2.1.0",
+    "redux-thunk": "2.2.0",
     "whatwg-fetch": "2.0.3"
   },
   "license": "MIT",
@@ -28,17 +29,18 @@
   },
   "devDependencies": {
     "clean-webpack-plugin": "0.1.17",
-    "es6-object-assign": "1.0.3",
+    "es6-object-assign": "1.1.0",
     "grunt": "^1.0.1",
     "grunt-contrib-watch": "^1.0.0",
     "grunt-watch-change": "^0.1.1",
-    "jest": "21.2.1",
+    "jest": "22.0.4",
     "jest-context": "^2.1.0",
     "jest-set": "^2.0.0",
-    "react-addons-test-utils": "15.3.2",
-    "sinon": "1.17.7",
-    "uglify-js": "3.1.3",
-    "webpack-dev-server": "2.9.1"
+    "react-addons-test-utils": "15.6.2",
+    "react-test-renderer": "^16.2.0",
+    "sinon": "4.1.3",
+    "uglify-js": "3.3.2",
+    "webpack-dev-server": "2.9.7"
   },
   "jest": {
     "testRegex": "(/test/.*|(\\_|/)spec)\\.js$",
