@@ -1,4 +1,5 @@
-import React, { Component, PropTypes } from 'react'
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import AddStopPoint from '../containers/AddStopPoint'
 import VisibleStopPoints from'../containers/VisibleStopPoints'
 import clone  from '../../helpers/clone'
@@ -16,8 +17,8 @@ export default class App extends Component {
         <VisibleStopPoints />
         <AddStopPoint />
       </div>
-    )  
-  } 
+    )
+  }
 }
 
 App.childContextTypes = {
