@@ -1,4 +1,6 @@
-import React, { PropTypes } from 'react'
+import React from 'react'
+import PropTypes from 'prop-types'
+
 import actions from '../actions'
 import TagsSelect2 from './TagsSelect2'
 
